refactor and add translations
