+++ shared-module/src/locales/en/quizzes.json
@@ -101,5 +101,8 @@
   "quiz-clickable-multiple-choice-name": "Select n",
   "quiz-clickable-multiple-choice-description": "Select n from presented buttons",
   "quiz-timeline-name": "Timeline",
-  "quiz-timeline-description": "Match years to events on a timeline"
+  "quiz-timeline-description": "Match years to events on a timeline",
+  "quiz-title-invalid-format": "Invalid format: check the tags",
+  "quiz-edit-title": "Edit",
+  "quiz-preview-title": "Preview"
 }
