+++ shared-module/src/locales/fi/quizzes.json
@@ -99,5 +99,8 @@
   "quiz-clickable-multiple-choice-name": "Valitse n",
   "quiz-clickable-multiple-choice-description": "Valitse n annetuista napeista",
   "quiz-timeline-name": "Aikajana",
-  "quiz-timeline-description": "Yhdistä vuodet tapahtumiin aikajanalla"
+  "quiz-timeline-description": "Yhdistä vuodet tapahtumiin aikajanalla",
+  "quiz-title-invalid-format": "Virheellinen muotoilu: tarkista tunnisteet",
+  "quiz-edit-title": "Muokkaa",
+  "quiz-preview-title": "Esikatsele"
 }
